Bugfix: Fix RDS triggers parameters so that they handle serialization/deserialization (#34222)
