+++ tests/providers/amazon/aws/triggers/test_ecs.py
@@ -27,8 +27,6 @@ from airflow import AirflowException
 from airflow.providers.amazon.aws.hooks.ecs import EcsHook
 from airflow.providers.amazon.aws.hooks.logs import AwsLogsHook
 from airflow.providers.amazon.aws.triggers.ecs import (
-    ClusterActiveTrigger,
-    ClusterInactiveTrigger,
     TaskDoneTrigger,
 )
 
@@ -97,36 +95,3 @@ class TestTaskDoneTrigger:
 
         assert response.payload["status"] == "success"
         assert response.payload["task_arn"] == "my_task_arn"
-
-
-class TestClusterTriggers:
-    @pytest.mark.parametrize(
-        "trigger",
-        [
-            ClusterActiveTrigger(
-                cluster_arn="my_arn",
-                aws_conn_id="my_conn",
-                waiter_delay=1,
-                waiter_max_attempts=2,
-                region_name="my_region",
-            ),
-            ClusterInactiveTrigger(
-                cluster_arn="my_arn",
-                aws_conn_id="my_conn",
-                waiter_delay=1,
-                waiter_max_attempts=2,
-                region_name="my_region",
-            ),
-        ],
-    )
-    def test_serialize_recreate(self, trigger):
-        class_path, args = trigger.serialize()
-
-        class_name = class_path.split(".")[-1]
-        clazz = globals()[class_name]
-        instance = clazz(**args)
-
-        class_path2, args2 = instance.serialize()
-
-        assert class_path == class_path2
-        assert args == args2
